Refactor puzzle wrapper
