+++ src/routes/[grid=gridkind]/[size=integer]/+page.svelte
@@ -1,13 +1,14 @@
 <script>
 	import { page } from '$app/stores';
-	import PuzzleWrapper from '$lib/puzzleWrapper/PuzzleWrapper.svelte';
+	import PuzzleKindWrapper from '$lib/puzzleWrapper/PuzzleKindWrapper.svelte';
 </script>
 
-<PuzzleWrapper
-	width={Number($page.params.size)}
-	height={Number($page.params.size)}
-	tiles={[]}
-	category={$page.params.grid}
-	size={Number($page.params.size)}
-	puzzleId={-1}
-/>
+{#key $page.url.pathname}
+	<PuzzleKindWrapper
+		width={Number($page.params.size)}
+		height={Number($page.params.size)}
+		category={$page.params.grid}
+		size={Number($page.params.size)}
+		puzzleId={-1}
+	/>
+{/key}
